Unify parse_identity function (phase 1)
